backend/requirements.txt
# =====================================================
# FASTAPI STACK
# =====================================================

fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart
email-validator

# =====================================================
# DATABASE
# =====================================================

sqlalchemy==2.0.49
psycopg2-binary
alembic

# =====================================================
# SECURITY
# =====================================================

python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# =====================================================
# ENV CONFIG
# =====================================================

python-dotenv
pydantic-settings

# =====================================================
# PAYMENTS
# =====================================================

razorpay==1.4.2

# =====================================================
# AI / OPENAI
# =====================================================

openai>=1.30.0

# =====================================================
# HTML GENERATION
# =====================================================

jinja2
pillow

# =====================================================
# CHARTS / ANALYTICS
# =====================================================

matplotlib
numpy

# =====================================================
# OPTIONAL UTILS
# =====================================================

python-dateutil
pypdf==4.2.0
playwright==1.48.0
